Move files to _models
