+++ benchmarks/_models/llama/eval.py
@@ -15,7 +15,7 @@ from generate import (
 from tokenizer import get_tokenizer
 
 import torchao
-from benchmarks._models.llama.model import prepare_inputs_for_model
+from torchao._models.model import prepare_inputs_for_model
 from torchao.quantization import (
     PerRow,
     PerTensor,
@@ -120,7 +120,7 @@ def run_evaluation(
             quantize_(model, int4_weight_only(layout=MarlinSparseLayout()))
         if "int4wo" in quantization and "gptq" in quantization:
             # avoid circular imports
-            from benchmarks._models._eval import MultiTensorInputRecorder
+            from torchao._models._eval import MultiTensorInputRecorder
             from torchao.quantization.GPTQ_MT import Int4WeightOnlyGPTQQuantizer
 
             groupsize = int(quantization.split("-")[-2])
@@ -172,7 +172,7 @@ def run_evaluation(
         if "autoround" in quantization:
             from transformers import AutoTokenizer
 
-            from benchmarks._models.llama.model import TransformerBlock
+            from torchao._models.model import TransformerBlock
             from torchao.prototype.autoround.autoround_llm import (
                 quantize_model_with_autoround_,
             )
@@ -242,7 +242,7 @@ def run_evaluation(
     with torch.no_grad():
         print("Running evaluation ...")
         # avoid circular imports
-        from benchmarks._models._eval import TransformerEvalWrapper
+        from torchao._models._eval import TransformerEvalWrapper
 
         TransformerEvalWrapper(
             model=model.to(device),
